requirements.txt
ccxt
python-decouple
pandas
numpy
ta
python-binance
fastquant
matplotlib==3.2.2
yfinance